requirements.txt
# Основные зависимости
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.0

# Open Deep Research
git+https://github.com/langchain-ai/open_deep_research.git

# Веб-фреймворк
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Поиск и данные
tavily-python>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# База данных
sqlalchemy>=2.0.0
alembic>=1.13.0

# Утилиты
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
loguru>=0.7.0

# Тестирование
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Документация
mkdocs>=1.5.0
mkdocs-material>=9.4.0
